Move BoundingBoxToPoint step to konduit-serving-pipeline
